Refactored the notebook structure.
